install() entrypoint now working for all builtin components via plugin system populated install() entrypoint for plugin system delete installers.py status_monitor plugin component working now
