+++ electrumsv-sdk/electrumsv_sdk/builtin_components/node/install.py
@@ -1,0 +1,5 @@
+import subprocess
+
+
+def fetch_node(app_state):
+    subprocess.run(f"{app_state.python} -m pip install electrumsv-node", shell=True, check=True)
