--- electrumsv-sdk/electrumsv_sdk/components/electrumsv/electrumsv.py
@@ -1,18 +1,0 @@
-def install(app_state):
-    pass
-
-
-def start(app_state):
-    pass
-
-
-def stop(app_state):
-    pass
-
-
-def reset(app_state):
-    pass
-
-
-def status_check(app_state):
-    pass
